Crate bholdus_tokens

Tokens Module

A simple, secure module for dealing with fungible tokens.

Overview

The Tokens module provides functionality for token managenment of fungible token classes with a fixed supply, including:

Token Issuance (Minting)

Token Transferal

Token Freezing

To use it in your runtime, you need to implement the tokens Config.

The supported dispatchable functions are documented in the Call enum.

Terminology

Admin: An account ID uniquely privileged to be able to unfreeze (thaw) an account and it’s assets

Token issuance/minting: The creation of a new token, whose total supply will belong to the account that issues the asset. This is a privileged operation.

Fungible token: An token whose units are interchangeable.

Issuer: An account ID uniquely privileged to be able to mint a particular class of assets.

Freezer: An account ID uniquely privileged to be able to freeze an account from transferring a particular class of tokens.

Freezing: Removing the possibility of an unpermissioned transfer of an token from a particular account.

Non-fungible token: An token for which each unit has unique characteristics.

Owner: An account ID uniquely privileged to be able to destroy a particular token class, or to set the Issuer, Freezer or Admin of that token class.

Approval: The act of allowing an account the permission to transfer some balance of token from approving account into some third-party destination account.

Sufficiency: The idea of a minimum-balance of an token being sufficient to allow the account’s existence on the system without requiring any other existential-deposit.

Interface

Permissionless Functions

create: Creates a new asset class, taking the required deposit.

transfer: Transfer sender’s assets to another account.

set_metadata: Set the metadata of an asset class.

clear_metadata: Remove the metadata of an asset class.

set_identity: Set the associated identity of an asset; a small deposit is reserved if not already taken.

clear_identity: Remove an asset’s associated identity; the deposit is returned.

Permissioned Functions

verify_asset: Verify the associated identity of an asset.

Privileged Functions

destroy: Destroys an entire asset class; called by the asset class’s Owner.

mint: Increases the asset balance of an account; called by the asset class’s Issuer.

freeze: Disallows further transfer’s from an accounts; called by the asset class’s Admin.

thaw: Allows further transfers from an account; called by the asset class’s Owner.

Public Functions

balance - Get the asset id balance of who.

total_supply - Get the total supply of an asset id.

Related Modules

System

Support

Re-exports

pub use pallet::*;

pub use weights::WeightInfo;

Modules

pallet
The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights
Autogenerated weights for pallet_assets

Structs

Approval
Data concering an approval.

AssetBalance

AssetDetails

AssetIdentity
Information concerning the identity of the controller of an account.

AssetMetadata

AssetProfile

BasicInformation
NOTE: This should be stored at the end of the storage item to facilitate the addition of extra fields in a backwards compatible way through a specialized Decod impl.

CurrencyAdapter

DestroyWitness
Witness data for the destroy transactions.

ExtraMutator
A mutator type allowing inspection and possible modification of the extra “sidecare” data.

NegativeImbalance

PositiveImbalance
Opaque, move-only struct with private fields that serves as a token denoting that funds have been created without any equal and opposite accounting.

Registration
Information concerning the identity of the controller of an account.

SocialProfile
NOTE: Social Profile

Enums

Data

Traits

FrozenBalance
Trait for allowing a minimum balance on the account to be specified, beyond the minimum_balance of the asset. This is additive - the minimum_balance of the asset must be met and then anything here in addition.

Type Definitions

BalanceOf

NegativeImbalanceOf