Struct bholdus_tokens::pallet::Pallet

pub struct Pallet<T, I = ()>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn adjust_extra(
id: T::AssetId,
who: impl Borrow<T::AccountId>
) -> Option<ExtraMutator<T, I>>

Return the extra “sid-car” data for id/who, or None if the account doesn’t exist.

pub fn total_issuance(id: T::AssetId) -> T::Balance

The total issuance of a token type.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn create(
origin: OriginFor<T>,
id: T::AssetId,
admin: <T::Lookup as StaticLookup>::Source,
min_balance: T::Balance
) -> DispatchResult

Issue a new class of fungible assets from a public origin.

This new asset class has no assets initially and its owner is the origin.

The origin must be Signed and the sender must have sufficient funds free.

Funds of sender are reserved by AssetDeposit.

Parameters:

id: The identifier of the new asset. This must not be currently in use to identify an existing asset.

admin: The admin of this class of assets. The admin is the initial address of each member of the asset class’s admin team

min_balance: The minimum balance of this new asset that any single account must have. If an account’s balance is reduced below this, then it collapses to zero.

Emits Created event when successful.

Weight: O(1)

pub fn force_create(
origin: OriginFor<T>,
id: T::AssetId,
owner: <T::Lookup as StaticLookup>::Source,
is_sufficient: bool,
min_balance: T::Balance
) -> DispatchResult

Issue a new class of fungible assets from a privileged origin.

This new asset class has no assets initially.

The origin must conform to ForceOrigin.

Unlike create, no funds are reserved.

id: The identifier of the new asset. This must not be currently in use to identify an existing asset.

owner: The owner of this class of assets. The owner has full superuser permissions. over this asset, but may later change and configure the permissions using transfer_ownership and set_team.

max_zombies: The total number of accounts which may hold assets in this class yet have no existential deposit.

min_balance: The minimum balance of this new asset that any single account must have. If an accounts balance is reduced below this, then it collapses to zero.

Emits ForceCreated event when successful.

Weight: O(1)

pub fn destroy(
origin: OriginFor<T>,
id: T::AssetId,
witness: DestroyWitness
) -> DispatchResult

Destroy a class of fungible assets.

The origin must conform to ForceOrigin or must be Signed and the sender must be the owner of the asset id.

id: The identifier of the asset to be destroyed. This must identity an existing asset.

Emits Destroyed event when successful.

Weight: O(c + p +a) where:

c = (witness.accounts - witness.sufficients)

s = witness.sufficients

a = witness.approvals

pub fn mint(
origin: OriginFor<T>,
id: T::AssetId,
beneficiary: <T::Lookup as StaticLookup>::Source,
amount: T::Balance
) -> DispatchResult

Mint assets of a particular class.

The origin must be Signed and the sender must be thi Issuer of the asset id.

id: The identifier of the asset to have some amount minted. beneficiary: The account to be credited with the minted assets. amount: The amount of the asset to be minted.

Emits Destroyed event when successful.

Weight: O(1) Modes: Pre-existing balance of beneficiary; Account pre-existence of beneficiary.

pub fn transfer(
origin: OriginFor<T>,
id: T::AssetId,
target: <T::Lookup as StaticLookup>::Source,
amount: T::Balance
) -> DispatchResult

Move some assets from the sender account to another.

Origin must be Signed.

id: The identifier of the asset to have some amount transferred.

target: The account to be credited.

amount: The amount by which the sender’s balance of assets should be reduced and target’s balance increase. The amount actually transferred may be slightly greather in the case that the transfer would otherwise take the sender balance above zero but below the minimum balance. Must be greather than zero.

Emits Transferred with the actual amount transferred. If this takes the source balance to below the minimum for the asset, then the amount transferred is increased to take it to zero.

Weight: O(1) Modes: Pre-existence of target; Post-existence of sender; Prior & post zombie-status of sender; Account pre-existence of target.

pub fn freeze(
origin: OriginFor<T>,
id: T::AssetId,
who: <T::Lookup as StaticLookup>::Source
) -> DispatchResult

Disallow further ungrivileged transfers from an account

Origin must be Signed and the sender should be the Freezer of the asset id.

id: The identifier of the asset to be frozen.

who: The account to be frozen.

Emits Frozen.

Weight: O(1)

pub fn thaw(
origin: OriginFor<T>,
id: T::AssetId,
who: <T::Lookup as StaticLookup>::Source
) -> DispatchResult

Allow unprivileged transfers from an account again.

Origin must be Signed and the sender should be the Admin of the asset id.

id: The identifier of the asset to be frozen.

who: The account to be unfrozen.

Emits Thawed

Weight: O(1)

pub fn freeze_asset(origin: OriginFor<T>, id: T::AssetId) -> DispatchResult

Disallow further ungrivileged transfers for the asset class.

Origin must be Signed and the sender should be the Freezer of the asset id.

id: The identifier of the asset to be frozen.

Emits Frozen.

Weight: O(1)

pub fn thaw_asset(origin: OriginFor<T>, id: T::AssetId) -> DispatchResult

Allow unprivileged transfers for the asset again.

Origin must be Signed and the sender should be the Admin of the asset id

id: The identifier of the asset to be frozen.

Emits Thawed.

Weight: O(1)

pub fn verify_asset(origin: OriginFor<T>, id: T::AssetId) -> DispatchResult

Verify asset from a privileged origin.

The origin must conform to ForceOrigin.

id: The identifier of the asset to verify.

Emits AssetVerified

Weight: O(1)

pub fn set_metadata(
origin: OriginFor<T>,
id: T::AssetId,
name: Vec<u8>,
symbol: Vec<u8>,
decimals: u8
) -> DispatchResult

Set the metadata for an asset.

Origin must be Signed and the sender should be the Owner of the asset id.

Funds of sender are reserved according to the formula: MetadataDepositBase + MetadataDepositPerByte * (name.len + symbol) taking into account any already reserved funds.

id: The identifier of the asset to update.

name: The user friendly name of this asset. Limited in length by StringLimit.

symbol: The exchange symbol for this asset. Limited in length by StringLimit.

decimals: The number of decimals this asset uses to represent on unit.

Emits MetadataSet

Weight: O(1)

pub fn clear_metadata(origin: OriginFor<T>, id: T::AssetId) -> DispatchResult

Clear the metadata for an asset.

Origin must be Signed and the sender should be the Owner of the asset id.

Any deposit is freed for the asset owner.

id: The identifier of the asset to clear.

Emits MetadataCleared.

Weight: O(1)

pub fn set_identity(
origin: OriginFor<T>,
id: T::AssetId,
info: AssetIdentity
) -> DispatchResultWithPostInfo

Set an account’s identity information and reserve the appropriate deposit.

If the asset already has identity information, the deposit is taken as part payment for new deposit.

The dispatch origin for this call must be Signed

info: The identity information. Emits IdentitySet if successful.

O(X+X' + R)

where X additional-field-count (deposit-bounded and code-bounded)

where R judgements-count (registrar-count-bounded)

One balance reserve operation.

One storage mutation (codec-read O(X' + R), codec-write O(X + R)).

One event. #

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn asset<KArg>(
k: KArg
) -> Option<AssetDetails<T::Balance, T::AccountId, <<T as Config<I>>::Currency as PalletCurrency<<T as SystemConfig>::AccountId>>::Balance>> where
KArg: EncodeLike<T::AssetId>,

Details of an asset.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn account<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> AssetBalance<T::Balance, T::Extra> where
KArg1: EncodeLike<T::AssetId>,
KArg2: EncodeLike<T::AccountId>,

The number of units of assets held by any given account.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn identity<KArg>(k: KArg) -> Option<Registration<BalanceOf<T, I>>> where
KArg: EncodeLike<T::AssetId>,

Information that is pertinet to identity the entity behind an account.

TWOX-NOTE: OK - AccountId is a secure hash.