Trait bholdus_tokens::pallet::Config

pub trait Config<I: 'static = ()>: Config {
    type Event: From<Event<Self, I>> + IsType<Self::Event>;
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen;
    type Amount: Signed + TryInto<Self::Balance> + TryFrom<Self::Balance> + Parameter + Member + SimpleArithmetic + Default + Copy + MaybeSerializeDeserialize;
    type AssetId: Member + Parameter + Copy + MaybeSerializeDeserialize + Ord;
    type Currency: PalletReservableCurrency<Self::AccountId>;
    type ForceOrigin: EnsureOrigin<Self::Origin>;
    type AssetDeposit: Get<<Self::Currency as PalletCurrency<Self::AccountId>>::Balance>;
    type BasicDeposit: Get<BalanceOf<Self, I>>;
    type FieldDeposit: Get<BalanceOf<Self, I>>;
    type MetadataDepositBase: Get<<Self::Currency as PalletCurrency<Self::AccountId>>::Balance>;
    type MetadataDepositPerByte: Get<<Self::Currency as PalletCurrency<Self::AccountId>>::Balance>;
    type ApprovalDeposit: Get<<Self::Currency as PalletCurrency<Self::AccountId>>::Balance>;
    type StringLimit: Get<u32>;
    type Freezer: FrozenBalance<Self::AssetId, Self::AccountId, Self::Balance>;
    type Extra: Member + Parameter + Default + MaxEncodedLen;
    type MaxAdditionalFields: Get<u32>;
    type MaxRegistrars: Get<u32>;
    type WeightInfo: WeightInfo;
    type ExistentialDeposits: GetByKey<Self::AssetId, Self::Balance>;
}

The module configuration trait.

Associated Types

type Event: From<Event<Self, I>> + IsType<Self::Event>

The overarching event type.

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen

The units in which we record balances.

type Amount: Signed + TryInto<Self::Balance> + TryFrom<Self::Balance> + Parameter + Member + SimpleArithmetic + Default + Copy + MaybeSerializeDeserialize

The amount type, should be signed version of Balance

type AssetId: Member + Parameter + Copy + MaybeSerializeDeserialize + Ord

Identifier for the class of asset.

type Currency: PalletReservableCurrency<Self::AccountId>

The currency mechanism.

type ForceOrigin: EnsureOrigin<Self::Origin>

The origin which may forcibly create or destroy an asset or otherwise alter privileged attributes.

type AssetDeposit: Get<<Self::Currency as PalletCurrency<Self::AccountId>>::Balance>

The basic amount of funds that must be reserved for an asset.

type BasicDeposit: Get<BalanceOf<Self, I>>

The amount held on deposit for a registred identiy

type FieldDeposit: Get<BalanceOf<Self, I>>

The amount held on deposit per additional field for a registered identity.

type MetadataDepositBase: Get<<Self::Currency as PalletCurrency<Self::AccountId>>::Balance>

The basic ammount of funds that must be reserved when adding metadata to your asset.

type MetadataDepositPerByte: Get<<Self::Currency as PalletCurrency<Self::AccountId>>::Balance>

The additional funds that must be reserved for the number of bytes you store in your metadata.

type ApprovalDeposit: Get<<Self::Currency as PalletCurrency<Self::AccountId>>::Balance>

The amount of funds that must be reserved when creating a new approval.

type StringLimit: Get<u32>

The maximum length of a name or symbol stored on-chain.

type Freezer: FrozenBalance<Self::AssetId, Self::AccountId, Self::Balance>

A hook to allow a per-asset, per-account minimum balance to be enforced. This must be respected in all permissionless operations.

type Extra: Member + Parameter + Default + MaxEncodedLen

Additional data to be stored with an account’s asset balance.

type MaxAdditionalFields: Get<u32>

Maximum number of additional fields that may be stored in an ID. Needed to bound the I/O required to access an identity, but can be pretty high.

type MaxRegistrars: Get<u32>

Maximum number of registrars allowed in the system. Needed to bound the complexity of, e.g., updating judgements.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type ExistentialDeposits: GetByKey<Self::AssetId, Self::Balance>

The minimum amount required to keep an account.

Loading content...